+++ src/test/direct_messenger/CMakeLists.txt
@@ -1,4 +1,5 @@
 # unittest_direct_messenger
-#add_executable(unittest_direct_messenger test_direct_messenger.cc DirectMessenger.cc)
+#add_library(QueueStrategy OBJECT QueueStrategy.cc)
+#add_executable(unittest_direct_messenger $<TARGET_OBJECTS:QueueStrategy> test_direct_messenger.cc DirectMessenger.cc)
 #add_ceph_unittest(unittest_direct_messenger)
 #target_link_libraries(unittest_direct_messenger global)
